code standart fix

+++ cpp/FileLoader.cpp
@@ -24,7 +24,7 @@ bool FileLoader::isEmpty() const
 int FileLoader::getNumber(const std::string& str, int& index) const
 {
 	std::string tmp;
-	while (str[index] != ' ' && index < static_cast<int>(str.length()))
+	while (str[index] != ' ' && index < str.length())
 	{
 		tmp.push_back(str[index++]);
 	}
@@ -41,7 +41,7 @@ std::vector<int> FileLoader::getNumbersSequence()
 
 	std::vector<int> tmpvec;		// вектор с ответом
 
-	for (int i = 0; i < static_cast<int>(str.length()); ++i)
+	for (int i = 0; i < str.length(); ++i)
 	{
 		if (str[i] != ' ')
 			tmpvec.push_back(getNumber(str, i));
